+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.x] - TBD
 ### Fixed
 - Fix some torchscriptability [#246]
+- Fix FourierMix being compatible with AMP [#258]
 
 ## [0.0.10] - 2022-03-14
 ### Fixed